Split debuginfo/line_info.rs from debuginfo/mod.rs
